refactor: pointcloud2_np读取的点云指定为float32

+++ ampcl/ros_utils/ros2_numpy/point_cloud2.py
@@ -252,7 +252,7 @@ def split_rgb_field(cloud_arr):
     return new_cloud_arr
 
 
-def get_xyz_points(cloud_array, remove_nans=True, dtype=np.float):
+def get_xyz_points(cloud_array, remove_nans=True, dtype=np.float32):
     '''Pulls out x, y, and z columns from the cloud recordarray, and returns
     a 3xN matrix.
     '''
@@ -277,7 +277,7 @@ def pointcloud2_to_xyz_array(cloud_msg, remove_nans=True):
         pointcloud2_to_array(cloud_msg), remove_nans=remove_nans)
 
 
-def get_xyzi_points(cloud_array, remove_nans=True, dtype=np.float):
+def get_xyzi_points(cloud_array, remove_nans=True, dtype=np.float32):
     '''Pulls out x, y, and z columns from the cloud recordarray, and returns
 	a 3xN matrix.
     '''
