+++ ampcl/ros_utils/ros_numpy/point_cloud2.py
@@ -229,7 +229,7 @@ def split_rgb_field(cloud_arr):
     return new_cloud_arr
 
 
-def get_xyz_points(cloud_array, remove_nans=True, dtype=np.float):
+def get_xyz_points(cloud_array, remove_nans=True, dtype=np.float32):
     '''Pulls out x, y, and z columns from the cloud recordarray, and returns
 	a 3xN matrix.
     '''
@@ -247,7 +247,7 @@ def get_xyz_points(cloud_array, remove_nans=True, dtype=np.float):
     return points
 
 
-def get_xyzi_points(cloud_array, remove_nans=True, dtype=np.float):
+def get_xyzi_points(cloud_array, remove_nans=True, dtype=np.float32):
     '''Pulls out x, y, and z columns from the cloud recordarray, and returns
 	a 3xN matrix.
     '''
